+++ CHANGELOG.md
@@ -42,6 +42,9 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Marker governance proposal are supported in cli [#367](https://github.com/provenance-io/provenance/issues/367)
 * Add ability to query metadata sessions by record [#212](https://github.com/provenance-io/provenance/issues/212)
 * Add Name and Symbol Cosmos features to Marker Metadata [#372](https://github.com/provenance-io/provenance/issues/372)
+* Add authz support to Marker module transfer `MarkerTransferAuthorization` [#265](https://github.com/provenance-io/provenance/issues/265)
+  * Add authz grant/revoke command to `marker` cli
+  * Add `granter` flag to `marker` transfer command to execute transfer with authz
 
 ### Bug Fixes